Add files via upload
BlackJack is a classic card game. It is designed here to be played against the computer. It has intermediate level coding and could be a good challenge for a new python programmer. It Uses classes ,functions, loops, try-excepts, lists, dictionaries conditional statements etc.
The classes here are added in a single file, but can also be made into different files and imported which can make the code more organized.